Add packages and scripts for connection to Neon PostgreSQL db
Both of these scripts are currently failing for me with the error
"password authentication failed for user [PostgreSQL role]" where
the role is NOT the role listed in my environment variable. I'm
troubleshooting that.

This change also does not add any of the code to actually hook up
the application to the database.